requirements_pytorch.txt
flask==2.3.3
ultralytics==8.0.196
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.0.1
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
torch==2.0.1
torchvision==0.15.2 